+++ sdk/apimanagement/arm-apimanagement/src/apiManagementClient.ts
@@ -351,7 +351,7 @@ export class ApiManagementClient extends coreClient.ServiceClient {
       credential: credentials,
     };
 
-    const packageDetails = `azsdk-js-arm-apimanagement/10.0.0`;
+    const packageDetails = `azsdk-js-arm-apimanagement/10.1.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -405,7 +405,7 @@ export class ApiManagementClient extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2024-05-01";
+    this.apiVersion = options.apiVersion || "2024-06-01-preview";
     this.apiGateway = new ApiGatewayImpl(this);
     this.apiManagementGatewaySkus = new ApiManagementGatewaySkusImpl(this);
     this.allPolicies = new AllPoliciesImpl(this);